+++ src/main/java/org/kuali/kra/subaward/bo/SubAwardAmountReleased.java
@@ -25,7 +25,7 @@ public class SubAwardAmountReleased  extends SubAwardAssociate {
     private static final long serialVersionUID = 1L;
 
     private Integer subAwardAmtReleasedId; 
-    private Integer subAwardId; 
+    private Long subAwardId; 
     private String subAwardCode; 
     private Integer lineNumber; 
     private KualiDecimal amountReleased; 
@@ -60,11 +60,11 @@ public class SubAwardAmountReleased  extends SubAwardAssociate {
         this.subAwardAmtReleasedId = subAwardAmtReleasedId;
     }
 
-    public Integer getSubAwardId() {
+    public Long getSubAwardId() {
         return subAwardId;
     }
 
-    public void setSubAwardId(Integer subAwardId) {
+    public void setSubAwardId(Long subAwardId) {
         this.subAwardId = subAwardId;
     }
